Use float to persist the total
- This avoids totals being returned as a string in tests, which is potentially surprising in context of examples

+++ db/migrate/20140425145318_create_sales.rb
@@ -1,7 +1,7 @@
 class CreateSales < ActiveRecord::Migration
   def change
     create_table :sales do |t|
-      t.decimal :total
+      t.float :total
       t.string :items
     end
   end
